+++ R/fusion_transforms.R
@@ -1107,7 +1107,7 @@ genFullInvFusionTransformMat <- function(
 ) {
 	##———— Safety checks ————————————————————————————————————————————
 	stopifnot(is.numeric(G), length(G) == 1L, G >= 1L)
-	stopifnot(is.numeric(T), length(T) == 1L, T >= 3L, G <= T - 1)
+	stopifnot(is.numeric(T), length(T) == 1L, T >= 2L, G <= T - 1)
 	stopifnot(is.numeric(d), length(d) == 1L, d >= 0L)
 	stopifnot(length(first_inds) == G)